Added Dca contract (#2)
* update readme

* changed location of deploy script

* refactor : changed location of deploy script

* added deploy script

* Landing page essentially done

* edited hardhat config

* Frontend

* ADDED A COIN

* Landing page more or less done

* Update README.md

* Linked to list view, still shit

* Frontend looks complete, yet to link upcuz not sure how

* Moved global variables to a context provider, updating of specific global variables not exactly confirmed yet

* Added some overlay, totally not styled

* updated README.md

* dcaVault

* dca vault

+++ packages/hardhat/contracts/mocks/Peanuts.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract Peanuts is ERC20 {
-    uint constant _initial_supply = 100 * (10**18);
+    uint256 constant _initial_supply = 100 * (10**18);
 
     constructor(string memory name, string memory symbol) ERC20(name, symbol) {
         _mint(msg.sender, _initial_supply);
